# Conflicts:
#	servicio-pruebas/pom.xml
#	servicio-pruebas/src/main/java/serviciopruebas/repositories/PruebaRepository.java
#	servicio-pruebas/src/main/java/serviciopruebas/services/PruebaService.java

+++ servicio-usuarios/src/main/java/serviciousuarios/repositories/InteresadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import serviciousuarios.entities.Interesado;
 
 @Repository
-public interface InteresadoRepository extends JpaRepository<Interesado, Long> {
+public interface InteresadoRepository extends JpaRepository<Interesado, Integer> {
 
 }
